Indexing metadata_0
Created a indexing python script which will index a given metadata file into a duckdb database, I then extract the information in the tables into a parquet which is then ready to use in ducklake. Adjustments to setup.py include updating the import to import the indexing data too, and updated the test function to print samples of each of the tables we have imported.